refactor: Service 및 Repository 리팩토링
- ReservationSerivce, ReservationTimeService 클래스가 의존하고 있는 ReservationRepository, ReservationTimeRepository의 인스턴스 변수 명을 일급 컬렉션처럼 변경.
- ReservationTimeRepository를 interface로 만들어 domain 패키지에 두고, 구현체의 이름을 JdbcReservationTimeRepository로 만들어 roomescape.infrastructure 패키지로 이동하여 도메인 레이어와 구분.
- ReservationRepsitory 구현체인 H2ReservationRepository의 클래스명을 JdbcReservationRepository로 변경하고, roomescape.infrastructure 패키지로 이동하여 도메인 레이어와 구분.
