+++ src/main/java/roomescape/application/ReservationTimeService.java
@@ -14,23 +14,23 @@ import roomescape.dto.ReservationTimeResponse;
 @Service
 @Transactional
 public class ReservationTimeService {
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimes;
+    private final ReservationRepository reservations;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository,
-                                  ReservationRepository reservationRepository) {
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.reservationRepository = reservationRepository;
+    public ReservationTimeService(ReservationTimeRepository reservationTimes,
+                                  ReservationRepository reservations) {
+        this.reservationTimes = reservationTimes;
+        this.reservations = reservations;
     }
 
     public ReservationTimeResponse create(ReservationTimeRequest request) {
-        ReservationTime reservationTime = reservationTimeRepository.create(request.from());
+        ReservationTime reservationTime = reservationTimes.create(request.from());
         return ReservationTimeResponse.from(reservationTime);
     }
 
     @Transactional(readOnly = true)
     public List<ReservationTimeResponse> findAll() {
-        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
+        List<ReservationTime> reservationTimes = this.reservationTimes.findAll();
         return convertToReservationTimeResponses(reservationTimes);
     }
 
@@ -41,15 +41,15 @@ public class ReservationTimeService {
     }
 
     public void deleteById(long id) {
-        Optional<ReservationTime> findReservationTime = reservationTimeRepository.findById(id);
+        Optional<ReservationTime> findReservationTime = reservationTimes.findById(id);
         if (findReservationTime.isEmpty()) {
             throw new IllegalArgumentException("존재하지 않는 예약 시간 입니다.");
         }
-        long reservedCount = reservationRepository.findReservationCountByTimeId(id);
+        long reservedCount = reservations.findReservationCountByTimeId(id);
         if (reservedCount > 0) {
             throw new IllegalStateException(String.format("해당 예약 시간에 연관된 예약이 존재하여 삭제할 수 없습니다. 삭제 요청한 시간:%s",
                     findReservationTime.get().getStartAt()));
         }
-        reservationTimeRepository.deleteById(id);
+        reservationTimes.deleteById(id);
     }
 }
